requirements.txt
Django==2.0.6
django-pure-pagination==0.3.0
Markdown==2.6.11
Pillow==5.1.0
Pygments==2.2.0
PyMySQL==0.8.1
pytz==2018.4
